pyproject.toml
[project]
name = "financial-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Bowen Li",email = "libowen085@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "plaid-python (>=37.1.0,<38.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "duckdb (>=1.4.2,<2.0.0)",
    "deltalake (>=1.2.1,<2.0.0)",
    "pyspark (>=4.0.1,<5.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "polars (>=1.35.2,<2.0.0)",
    "delta-spark (>=4.0.0,<5.0.0)",
    "jsonschema (>=4.25.1,<5.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "financial_tracker", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.5,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pre-commit (>=4.4.0,<5.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "types-jsonschema (>=4.25.1.20251009,<5.0.0.0)"
]
